Refactoring: Move structs into modules
If they all live in the main.rs, this is going to grow quickly as more
pin types get introduced.
